Modernize notifications redux
