+++ src/redux/actions/index.ts
@@ -1,6 +1,5 @@
 import firebase from 'firebase';
 import { helperActions } from './helper';
-import { notificationsActions } from './notifications';
 import { previousSpeakersActions } from './previous-speakers';
 import { scheduleActions } from './schedule';
 import { sessionsActions } from './sessions';
@@ -9,7 +8,6 @@ import { userActions } from './user';
 
 export {
   helperActions,
-  notificationsActions,
   previousSpeakersActions,
   scheduleActions,
   sessionsActions,
