+++ src/redux/notifications/state.ts
@@ -1,0 +1,5 @@
+import { NotificationState, NOTIFICATIONS_STATUS } from './types';
+
+export const initialNotificationState: NotificationState = {
+  status: NOTIFICATIONS_STATUS.DEFAULT,
+};
